backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
defusedxml==0.7.1
Django==5.0
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
flake8==6.1.0
flake8-import-order==0.18.2
setuptools==65.5.0
idna==3.10
isort==5.0.0
mccabe==0.7.0
oauthlib==3.2.2
pillow==11.2.1
psycopg2-binary==2.9.10
pycodestyle==2.11.0
pycparser==2.22
pyflakes==3.1.0
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
